requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
async-property==0.2.2
async-timeout==5.0.1
attrs==24.2.0
Authlib==1.3.2
bitarray==3.0.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cryptography==43.0.3
cytoolz==1.0.0
deprecation==2.1.0
dill==0.3.9
eth-account==0.13.7
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
exceptiongroup==1.2.2
Flask==3.0.3
flask-cors==5.0.1
flask-oidc==2.2.2
flask-restx==1.3.0
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
packaging==23.2
parsimonious==0.10.0
pi==0.1.2
platformdirs==4.3.7
propcache==0.2.1
py-solc-x==2.0.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pylint==3.3.6
pylint-flask==0.6
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
python-keycloak==5.3.1
pytz==2024.2
pyunormalize==16.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
rpds-py==0.20.1
six==1.16.0
sniffio==1.3.1
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
web3==7.13.0
websockets==13.1
Werkzeug==3.1.1
yarl==1.18.3
